refactor: melhora a logica das services e cria mais mensagens de excessão
