+++ src/main/java/com/web/desenvolvimento/edusphere/controllers/AuthenticationController.java
@@ -3,8 +3,8 @@ package com.web.desenvolvimento.edusphere.controllers;
 import com.web.desenvolvimento.edusphere.domain.user.Role;
 import com.web.desenvolvimento.edusphere.domain.user.User;
 import com.web.desenvolvimento.edusphere.domain.user.exceptions.UserNotFoundException;
-import com.web.desenvolvimento.edusphere.dto.LoginRequest;
-import com.web.desenvolvimento.edusphere.dto.LoginResponse;
+import com.web.desenvolvimento.edusphere.dto.login.LoginRequest;
+import com.web.desenvolvimento.edusphere.dto.login.LoginResponse;
 import com.web.desenvolvimento.edusphere.dto.user.UserRequestDTO;
 import com.web.desenvolvimento.edusphere.mappers.IUserMapper;
 import com.web.desenvolvimento.edusphere.repositories.IUserRepository;
