+++ src/main/java/com/web/desenvolvimento/edusphere/domain/manager/expections/ManagerNotFoundException.java
@@ -1,0 +1,7 @@
+package com.web.desenvolvimento.edusphere.domain.manager.expections;
+
+public class ManagerNotFoundException extends RuntimeException {
+    public ManagerNotFoundException(String message) {
+        super(message);
+    }
+}
