+++ src/main/java/com/web/desenvolvimento/edusphere/domain/teacher/exceptions/DepartmentOrUserNotFoundException.java
@@ -1,0 +1,7 @@
+package com.web.desenvolvimento.edusphere.domain.teacher.exceptions;
+
+public class DepartmentOrUserNotFoundException extends RuntimeException{
+    public DepartmentOrUserNotFoundException(String message) {
+        super(message);
+    }
+}
