+++ src/main/java/com/web/desenvolvimento/edusphere/dto/login/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.web.desenvolvimento.edusphere.dto;
+package com.web.desenvolvimento.edusphere.dto.login;
 
 public record LoginRequest(String username, String password){
 }
