+++ src/main/java/com/web/desenvolvimento/edusphere/dto/login/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.web.desenvolvimento.edusphere.dto;
+package com.web.desenvolvimento.edusphere.dto.login;
 
 public record LoginResponse(String accessToken, String username, String role) {
 }
